// apps/web/src/api/generated/models/grantCreditsDtoCreditType.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Script Factory API
 * API for UGC Script Factory
 * OpenAPI spec version: 1.0
 */

export type GrantCreditsDtoCreditType = typeof GrantCreditsDtoCreditType[keyof typeof GrantCreditsDtoCreditType];


export const GrantCreditsDtoCreditType = {
  free: 'free',
  subscription: 'subscription',
  pack: 'pack',
} as const;